all component'd out.
